feat: getActiveQuoteMetaOrDefault
