+++ src/state/slices/tradeQuoteSlice/tradeQuoteSlice.ts
@@ -1,7 +1,6 @@
 import type { PayloadAction } from '@reduxjs/toolkit'
 import { createSlice } from '@reduxjs/toolkit'
 import type { SwapperName, TradeQuote } from '@shapeshiftoss/swapper'
-import type { PartialRecord } from '@shapeshiftoss/types'
 import { orderBy, uniqBy } from 'lodash'
 import type { ApiQuote } from 'state/apis/swapper'
 
@@ -9,20 +8,9 @@ import { initialState, initialTradeExecutionState } from './constants'
 import {
   HopExecutionState,
   type StreamingSwapMetadata,
-  type TradeExecutionMetadata,
   TradeExecutionState,
   TransactionExecutionState,
 } from './types'
-
-export type TradeQuoteSliceState = {
-  activeStep: number | undefined // Make sure to actively check for undefined vs. falsy here. 0 is the first step, undefined means no active step yet
-  activeQuoteMeta: { swapperName: SwapperName; identifier: string } | undefined // the selected quote metadata used to find the active quote in the api responses
-  confirmedQuote: TradeQuote | undefined // the quote being executed
-  tradeExecution: TradeExecutionMetadata
-  tradeQuotes: PartialRecord<SwapperName, Record<string, ApiQuote>> // mapping from swapperName to quoteId to ApiQuote
-  tradeQuoteDisplayCache: ApiQuote[]
-  isTradeQuoteRequestAborted: boolean // used to conditionally render results and loading state
-}
 
 export const tradeQuoteSlice = createSlice({
   name: 'tradeQuote',
